+++ bootique-jersey-jakarta-client-junit5-wiremock/src/test/java/io/bootique/jersey/client/junit5/wiremock/junit/TestWithEmulatedBackend.java
@@ -22,20 +22,31 @@ import io.bootique.BQCoreModule;
 import io.bootique.BQRuntime;
 import io.bootique.Bootique;
 import io.bootique.jersey.JerseyModule;
+import io.bootique.jetty.JettyModule;
 import io.bootique.junit5.BQApp;
 import io.bootique.junit5.BQTest;
 import io.bootique.junit5.BQTestScope;
+import io.bootique.junit5.BQTestTool;
+import io.bootique.junit5.scope.BQBeforeMethodCallback;
+import jakarta.servlet.*;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.QueryParam;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+import org.junit.jupiter.api.extension.ExtensionContext;
+
+import java.io.IOException;
+import java.util.concurrent.atomic.AtomicInteger;
 
 @BQTest
 public abstract class TestWithEmulatedBackend {
 
     protected static final String SERVER_URL = "http://localhost:16348";
+
+    @BQTestTool(BQTestScope.GLOBAL)
+    protected static RequestCounter counter = new RequestCounter();
 
     @BQApp(BQTestScope.GLOBAL)
     protected static BQRuntime server = Bootique.app("--server")
@@ -46,6 +57,7 @@ public abstract class TestWithEmulatedBackend {
             .module(b -> BQCoreModule.extend(b).setProperty("bq.jetty.connectors[0].port", "16348"))
 
             .module(b -> JerseyModule.extend(b).addResource(Resource.class))
+            .module(b -> JettyModule.extend(b).addFilter(counter, "counter", 0, "/*"))
             .createRuntime();
 
     @Path("")
@@ -68,6 +80,26 @@ public abstract class TestWithEmulatedBackend {
         @Path("/p1/p11")
         public Response getSubPath() {
             return Response.ok().entity("get:p1:p11").build();
+        }
+    }
+
+    protected static int getMethodRequestCount() {
+        return counter.counter.get();
+    }
+
+    public static class RequestCounter implements Filter, BQBeforeMethodCallback {
+
+        static final AtomicInteger counter = new AtomicInteger();
+
+        @Override
+        public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws ServletException, IOException {
+            counter.addAndGet(1);
+            chain.doFilter(request, response);
+        }
+
+        @Override
+        public void beforeMethod(BQTestScope scope, ExtensionContext context) throws Exception {
+            counter.set(0);
         }
     }
 }